moved process_batch_args to utils
